feat(k3): paged MLA latent KV pool + write path (M1)
Add the paged latent cache the K3 MLA layers will decode from: one
page-first slab per state pool ([page][mla_layer][64 tokens][576] bf16 —
the post-norm kv latent and the shared rope half, the exact quantity
kv_b expands K and V from; NoPE, so the row is position-independent),
a per-slot block table with a host mirror, and a plain free-list page
allocator (claim on 64-boundary crossing, release-all on slot reset,
whole-page copy on prefill adoption). Pool size comes from the new
K3ExecutorConfig::kv_pages (0 derives full coverage: every slot can
reach max_ctx).

The decode step now appends every MLA layer's new-token latent into the
current page slice (two indexed writes through the per-layer row shift)
alongside the existing expanded K/V append. This is a transitional dual
write: attention still reads the expanded cache, so behavior is
unchanged and every gate stays green; M2 lands the absorbed-MLA kernel
over the pages and deletes the expanded cache and its kernel family.

The shared indexed row write widens its element index to size_t — the
K3 paged slab is the first destination past 2^31 elements.

Gates: cargo test -p pegainfer-k3 --lib (34 passed) and cargo check
--tests green; GPU gates unaffected by construction (attention path
untouched).

+++ pegainfer-kernels/csrc/shared/elementwise.cu
@@ -177,7 +177,10 @@ __global__ void scaled_add_rows_indexed_kernel(
          row < rows;
          row += gridDim.x * blockDim.x) {
       int delta_idx = token * rows + row;
-      int out_idx = out_token * out_hidden_dim + row_offset + row;
+      // out_token * out_hidden_dim overflows i32 for large row pools (the K3
+      // paged-KV slab exceeds 2^31 elements), so the element index is size_t.
+      size_t out_idx =
+          (size_t)out_token * out_hidden_dim + row_offset + row;
       float base = __bfloat162float(out[out_idx]);
       float add = __bfloat162float(delta[delta_idx]) * scale;
       out[out_idx] = __float2bfloat16(base + add);
